+++ src/main/java/kotlinx/reflect/lite/CallableMetadata.kt
@@ -19,7 +19,7 @@ package kotlinx.reflect.lite
 /**
  * Provides access to the metadata of callable Kotlin declarations: functions, properties and constructors.
  */
-interface CallableMetadata {
+interface CallableMetadata : DeclarationMetadata {
     /**
      * Metadata for the parameters of this callable declaration, in the declaration order.
      * If this declaration is a property, an empty list is returned.