+++ src/main/java/kotlinx/reflect/lite/impl/util.kt
@@ -16,6 +16,8 @@
 
 package kotlinx.reflect.lite.impl
 
+import kotlinx.reflect.lite.DeclarationMetadata
+import org.jetbrains.kotlin.serialization.ProtoBuf
 import java.lang.ref.SoftReference
 import java.lang.reflect.Array
 import kotlin.reflect.KProperty
@@ -37,3 +39,12 @@ internal fun Class<*>.desc(): String {
     if (this == Void.TYPE) return "V"
     return Array.newInstance(this, 0).javaClass.name.substring(1).replace('.', '/')
 }
+
+internal val ProtoBuf.Visibility.toVisibility: DeclarationMetadata.Visibility?
+    get() = when (this) {
+        ProtoBuf.Visibility.PUBLIC -> DeclarationMetadata.Visibility.PUBLIC
+        ProtoBuf.Visibility.PROTECTED -> DeclarationMetadata.Visibility.PROTECTED
+        ProtoBuf.Visibility.INTERNAL -> DeclarationMetadata.Visibility.INTERNAL
+        ProtoBuf.Visibility.PRIVATE, ProtoBuf.Visibility.PRIVATE_TO_THIS -> DeclarationMetadata.Visibility.PRIVATE
+        else -> null
+    }