+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add ClassMetadata.isData
 * Add ClassMetadata.kind
 * Add ConstructorMetadata.isPrimary
+* Add DeclarationMetadata.visibility
 * Update to Kotlin 1.2.0
 
 ## 1.0.0 (Feb 15, 2016)